refactor(TripTrackingData): Move helper methods from ManageTripTracking.
